libopencash: add support for android
There is an error remaining when calling vector.erase(). I have not yet
figured out the root cause, but compiling goes through when commenting
out line 84 in src/model/Account.cpp:

    // children.erase(child);

This is a quick fix to unblock further progress on the android
infrastructure. Obviously it isn't the long-term fix.

Here is the error message:

.../opencash/libopencash/src/model/Account.cpp:84:18: error: no matching member
      function for call to 'erase'
        children.erase(child);
        ~~~~~~~~~^~~~~
/opt/android/android-ndks/standalone/arm-linux-androideabi-clang3.3-android-9/include/c++/4.6/bits/stl_vector.h:1047:7: note:
      candidate function not viable: no known conversion from '__normal_iterator<const class
      std::weak_ptr<class opencash::model::Account> *, [...]>' to '__normal_iterator<pointer,
      [...]>' for 1st argument
      erase(iterator __position);